refactor: root page parallel routes 적용
