+++ app/(root)/page.tsx
@@ -1,0 +1,5 @@
+const Home = async () => {
+  return <></>;
+};
+
+export default Home;
