+++ app/default.tsx
@@ -1,0 +1,5 @@
+export const Default = () => {
+  return null;
+};
+
+export default Default;
